move functions out of Context.Traverse.Types
