+++ library/Translate/src/Juvix/Core/Common/Context/Traverse/Types.hs
@@ -7,7 +7,7 @@ module Juvix.Core.Common.Context.Traverse.Types
     Group',
     Groups,
     Groups',
-    Prefix,
+    Prefix (..),
     Deps,
 
     -- * Capabilities
@@ -19,22 +19,12 @@ module Juvix.Core.Common.Context.Traverse.Types
     HasRecGroups,
     run,
     run_,
-
-    -- * Operations
-    addGroup,
-    withPrefix,
-    qualify,
-    prefixM,
-    applyPrefix,
-    addDeps,
   )
 where
 
 import qualified Data.DList as D
 import Data.HashMap.Strict (HashMap)
-import qualified Data.HashMap.Strict as HashMap
 import Data.HashSet (HashSet)
-import qualified Data.HashSet as HashSet
 import Juvix.Core.Common.Context.Types
 import qualified Juvix.Core.Common.Context.Types as Context
 import Juvix.Library
@@ -138,47 +128,3 @@ run curNameSpace (Env act) =
   where
     initState = S {prefix, output = [], curNameSpace, deps = []}
     prefix = P [Context.topLevelName]
-
--- | Add a group to the final output.
-addGroup ::
-  (PrefixReader m, OutputState term ty sumRep m, Foldable t) =>
-  t (Entry term ty sumRep) ->
-  m ()
-addGroup grp = do
-  prefix <- prefixM
-  case nonEmpty $ toList grp of
-    Just grp -> modify @"output" $ HashMap.alter f prefix
-      where
-        f = Just . maybe [grp] (<> [grp])
-    Nothing -> pure ()
-
--- | Add dependencies on the given names to the current namespace.
-addDeps :: (Foldable t, DepsState m, PrefixReader m) => t NameSymbol.T -> m ()
-addDeps deps = do
-  let mods = HashSet.fromList $ map NameSymbol.mod $ toList deps
-  let f = Just . maybe mods (HashSet.union mods)
-  prefix <- prefixM
-  modify @"deps" $ HashMap.alter f prefix
-
-toMod :: Prefix -> NameSymbol.Mod
-toMod (P p) = toList p
-
-prefixM :: PrefixReader m => m NameSymbol.Mod
-prefixM = asks @"prefix" toMod
-
--- | Extend the current module prefix.
---
--- >>> 'fst' $ 'run' $ 'withPrefix' \"A\" $ 'qualify' \"X\"
--- A.X
--- >>> 'fst' $ 'run' $ 'withPrefix' \"A\" $ 'withPrefix' \"B\" $ 'qualify' \"X\"
--- A.B.X
-withPrefix :: PrefixReader m => Symbol -> m a -> m a
-withPrefix n = local @"prefix" \(P pfx) -> P $ D.snoc pfx n
-
--- | Qualify a name by the current module prefix.
-qualify :: PrefixReader m => Symbol -> m NameSymbol.T
-qualify n = asks @"prefix" \pfx -> applyPrefix pfx n
-
--- | Apply a prefix to a name.
-applyPrefix :: Prefix -> Symbol -> NameSymbol.T
-applyPrefix (P pfx) = NameSymbol.qualify1 pfx
